backend/pyproject.toml
[project]
name = "cerebro-osrs-backend"
version = "0.1.0"
description = "FastAPI backend for the Cerebro OSRS monorepo."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.0",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
asyncio_mode = "auto"
testpaths = ["tests"]
